requirements.txt
alembic==1.6.5
aniso8601==8.0.0
certifi==2021.5.30
charset-normalizer==2.0.1
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
gunicorn==20.1.0
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.12.2
psycopg2-binary==2.8.6
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2020.1
requests==2.26.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.26.6
Werkzeug==2.0.1